+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.2.0]
 
 - Added file system watcher to monitor for changes to the current document.
   This allows the preview to be automatically updated whenever the ZPL file is changed.